fix build warnings

+++ src/tkc/mem_pool.h
@@ -131,10 +131,10 @@ static inline ret_t mem_pool_put_index(mem_pool_t* pool, uint32_t index) {
 static inline int32_t mem_pool_get_index(mem_pool_t* pool, void* ptr) {
   uint8_t* addr = (uint8_t*)ptr;
   if (pool->start <= addr && addr < (pool->start + pool->block_size * pool->block_nr)) {
-    uint32_t offset = addr - pool->start;
-    uint32_t index = offset / (pool->block_size);
+    int64_t offset = addr - pool->start;
+    int64_t index = offset / (pool->block_size);
     assert(offset % pool->block_size == 0);
-    return index;
+    return (int32_t)index;
   }
 
   return -1;
